Accurate autonomous driving (for real this time)
I tested different distances and multiplied them by proporitons to
reverse engineer the gear ratio, because we couldn't find an accurate
one from the website or looking at the gears. I then realized it was
overshooting, so I implimented my own jank PD loop to prevent that. Now
the drive to distance command doesn't take a throttle, but it takes a
direction (1 for forewards, -1 for backwards).