Cleaning our SBPLI code (part 1)
I created this branch for the case that we want to share or publish the
code we used at SBPLI. I will be commiting to clean the code without
changing functionality. In this commit I removed code that was not used
on our SBPLI robot, from autos, to commands, to subsystems, and I
resolved TODOs remaining in the code.